+++ test/ExecutionFramework.cpp
@@ -101,7 +101,9 @@ u256 ExecutionFramework::gasPrice() const
 
 u256 ExecutionFramework::blockHash(u256 const& _number) const
 {
-	return {EVMHost::convertFromEVMC(m_evmHost->get_block_hash(uint64_t(_number & numeric_limits<uint64_t>::max())))};
+	return {EVMHost::convertFromEVMC(
+		m_evmHost->get_block_hash(static_cast<int64_t>(_number & numeric_limits<uint64_t>::max()))
+	)};
 }
 
 u256 ExecutionFramework::blockNumber() const
@@ -153,7 +155,7 @@ void ExecutionFramework::sendMessage(bytes const& _data, bool _isCreation, u256 
 	if (m_showMessages)
 	{
 		cout << " out:     " << toHex(m_output) << endl;
-		cout << " result: " << size_t(result.status_code) << endl;
+		cout << " result: " << static_cast<size_t>(result.status_code) << endl;
 		cout << " gas used: " << m_gasUsed.str() << endl;
 	}
 }
@@ -180,7 +182,7 @@ void ExecutionFramework::sendEther(Address const& _addr, u256 const& _amount)
 
 size_t ExecutionFramework::currentTimestamp()
 {
-	return m_evmHost->tx_context.block_timestamp;
+	return static_cast<size_t>(m_evmHost->tx_context.block_timestamp);
 }
 
 size_t ExecutionFramework::blockTimestamp(u256 _block)
@@ -188,7 +190,7 @@ size_t ExecutionFramework::blockTimestamp(u256 _block)
 	if (_block > blockNumber())
 		return 0;
 	else
-		return size_t((currentTimestamp() / blockNumber()) * _block);
+		return static_cast<size_t>((currentTimestamp() / blockNumber()) * _block);
 }
 
 Address ExecutionFramework::account(size_t _idx)
